--- a/lib/storage/WebStorage.js
+++ b/lib/storage/WebStorage.js
@@ -4,7 +4,7 @@
  * data changes and then stay up-to-date with everything happening in Onyx.
  */
 import _ from 'underscore';
-import Storage from './providers/LocalForage';
+import Storage from './providers/IDBKeyVal';
 
 const SYNC_ONYX = 'SYNC_ONYX';