--- a/lib/storage/WebStorage.js
+++ b/lib/storage/WebStorage.js
@@ -1,5 +1,5 @@
 /**
- * This file is here to wrap LocalForage with a layer that provides data-changed events like the ones that exist
+ * This file is here to wrap IDBKeyVal with a layer that provides data-changed events like the ones that exist
  * when using LocalStorage APIs in the browser. These events are great because multiple tabs can listen for when
  * data changes and then stay up-to-date with everything happening in Onyx.
  */