--- a/lib/storage/__mocks__/index.js
+++ b/lib/storage/__mocks__/index.js
@@ -8,7 +8,7 @@ const set = jest.fn((key, value) => {
     return Promise.resolve(value);
 });
 
-const localForageMock = {
+const idbKeyvalMock = {
     setItem(key, value) {
         return set(key, value);
     },
@@ -57,23 +57,23 @@ const localForageMock = {
     config() {},
 };
 
-const localForageMockSpy = {
-    localForageSet: set,
-    setItem: jest.fn(localForageMock.setItem),
-    getItem: jest.fn(localForageMock.getItem),
-    removeItem: jest.fn(localForageMock.removeItem),
-    removeItems: jest.fn(localForageMock.removeItems),
-    clear: jest.fn(localForageMock.clear),
-    getAllKeys: jest.fn(localForageMock.getAllKeys),
-    config: jest.fn(localForageMock.config),
-    multiGet: jest.fn(localForageMock.multiGet),
-    multiSet: jest.fn(localForageMock.multiSet),
-    multiMerge: jest.fn(localForageMock.multiMerge),
-    mergeItem: jest.fn(localForageMock.mergeItem),
+const idbKeyvalMockSpy = {
+    idbKeyvalSet: set,
+    setItem: jest.fn(idbKeyvalMock.setItem),
+    getItem: jest.fn(idbKeyvalMock.getItem),
+    removeItem: jest.fn(idbKeyvalMock.removeItem),
+    removeItems: jest.fn(idbKeyvalMock.removeItems),
+    clear: jest.fn(idbKeyvalMock.clear),
+    getAllKeys: jest.fn(idbKeyvalMock.getAllKeys),
+    config: jest.fn(idbKeyvalMock.config),
+    multiGet: jest.fn(idbKeyvalMock.multiGet),
+    multiSet: jest.fn(idbKeyvalMock.multiSet),
+    multiMerge: jest.fn(idbKeyvalMock.multiMerge),
+    mergeItem: jest.fn(idbKeyvalMock.mergeItem),
     getStorageMap: jest.fn(() => storageMapInternal),
     setInitialMockData: jest.fn((data) => {
         storageMapInternal = data;
     }),
 };
 
-export default localForageMockSpy;
+export default idbKeyvalMockSpy;