--- a/lib/storage/providers/IDBKeyVal.js
+++ b/lib/storage/providers/IDBKeyVal.js
@@ -13,7 +13,6 @@ import {
 import _ from 'underscore';
 import fastMerge from '../../fastMerge';
 
-// Same config as localforage, so we can swap the providers easily
 const customStore = createStore('OnyxDB', 'keyvaluepairs');
 
 const provider = {