--- a/package.json
+++ b/package.json
@@ -57,6 +57,7 @@
     "eslint-config-expensify": "^2.0.38",
     "eslint-plugin-jsx-a11y": "^6.6.1",
     "eslint-plugin-react": "^7.31.10",
+    "idb-keyval": "^6.2.1",
     "jest": "^26.5.2",
     "jest-cli": "^26.5.2",
     "jsdoc-to-markdown": "^7.1.0",
@@ -73,6 +74,7 @@
     "webpack-merge": "^5.8.0"
   },
   "peerDependencies": {
+    "idb-keyval": "^6.2.1",
     "localforage": "^1.10.0",
     "localforage-removeitems": "^1.4.0",
     "react": ">=18.1.0",
@@ -80,6 +82,9 @@
     "react-native-quick-sqlite": "^8.0.0-beta.2"
   },
   "peerDependenciesMeta": {
+    "idb-keyval": {
+      "optional": true
+    },
     "react-native-performance": {
       "optional": true
     },