--- a/package.json
+++ b/package.json
@@ -61,7 +61,6 @@
     "jest": "^26.5.2",
     "jest-cli": "^26.5.2",
     "jsdoc-to-markdown": "^7.1.0",
-    "localforage": "^1.10.0",
     "metro-react-native-babel-preset": "^0.72.3",
     "prop-types": "^15.7.2",
     "react": "18.2.0",
@@ -75,8 +74,6 @@
   },
   "peerDependencies": {
     "idb-keyval": "^6.2.1",
-    "localforage": "^1.10.0",
-    "localforage-removeitems": "^1.4.0",
     "react": ">=18.1.0",
     "react-native-performance": "^4.0.0",
     "react-native-quick-sqlite": "^8.0.0-beta.2"
@@ -89,12 +86,6 @@
       "optional": true
     },
     "react-native-quick-sqlite": {
-      "optional": true
-    },
-    "localforage": {
-      "optional": true
-    },
-    "localforage-removeitems": {
       "optional": true
     }
   },