--- a/tests/unit/storage/providers/IDBKeyvalProviderTest.js
+++ b/tests/unit/storage/providers/IDBKeyvalProviderTest.js
@@ -1,10 +1,10 @@
 import _ from 'underscore';
 
-import LocalForageProviderMock from '../../../../lib/storage/providers/LocalForage';
+import IDBKeyValProviderMock from '../../../../lib/storage/providers/IDBKeyVal';
 import createDeferredTask from '../../../../lib/createDeferredTask';
 import waitForPromisesToResolve from '../../../utils/waitForPromisesToResolve';
 
-describe('storage/providers/LocalForage', () => {
+describe('storage/providers/IDBKeyVal', () => {
     const SAMPLE_ITEMS = [
         ['string', 'Plain String'],
         ['array', ['Mixed', {array: [{id: 1}, {id: 2}]}]],
@@ -18,8 +18,8 @@ describe('storage/providers/LocalForage', () => {
     beforeAll(() => jest.useRealTimers());
     beforeEach(() => {
         jest.clearAllMocks();
-        LocalForageProviderMock.clear();
-        LocalForageProviderMock.clear.mockClear();
+        IDBKeyValProviderMock.clear();
+        IDBKeyValProviderMock.clear.mockClear();
     });
 
     it('multiSet', () => {
@@ -27,20 +27,20 @@ describe('storage/providers/LocalForage', () => {
         const pairs = SAMPLE_ITEMS.slice();
 
         // When they are saved
-        return LocalForageProviderMock.multiSet(pairs).then(() => {
-            // We expect a call to localForage.setItem for each pair
-            _.each(pairs, ([key, value]) => expect(LocalForageProviderMock.setItem).toHaveBeenCalledWith(key, value));
+        return IDBKeyValProviderMock.multiSet(pairs).then(() => {
+            // We expect a call to idbKeyval.setItem for each pair
+            _.each(pairs, ([key, value]) => expect(IDBKeyValProviderMock.setItem).toHaveBeenCalledWith(key, value));
         });
     });
 
     it('multiGet', () => {
         // Given we have some data in storage
-        LocalForageProviderMock.multiSet(SAMPLE_ITEMS);
+        IDBKeyValProviderMock.multiSet(SAMPLE_ITEMS);
 
         return waitForPromisesToResolve().then(() => {
         // Then multi get should retrieve them
             const keys = _.map(SAMPLE_ITEMS, _.head);
-            return LocalForageProviderMock.multiGet(keys)
+            return IDBKeyValProviderMock.multiGet(keys)
                 .then(pairs => expect(pairs).toEqual(expect.arrayContaining(SAMPLE_ITEMS)));
         });
     });
@@ -59,10 +59,10 @@ describe('storage/providers/LocalForage', () => {
             traits: {hair: 'black'},
         };
 
-        LocalForageProviderMock.multiSet([['@USER_1', USER_1], ['@USER_2', USER_2]]);
+        IDBKeyValProviderMock.multiSet([['@USER_1', USER_1], ['@USER_2', USER_2]]);
 
         return waitForPromisesToResolve().then(() => {
-            LocalForageProviderMock.localForageSet.mockClear();
+            IDBKeyValProviderMock.idbKeyvalSet.mockClear();
 
             // Given deltas matching existing structure
             const USER_1_DELTA = {
@@ -76,12 +76,12 @@ describe('storage/providers/LocalForage', () => {
             };
 
             // When data is merged to storage
-            return LocalForageProviderMock.multiMerge([
+            return IDBKeyValProviderMock.multiMerge([
                 ['@USER_1', USER_1_DELTA],
                 ['@USER_2', USER_2_DELTA],
             ]).then(() => {
                 // Then each existing item should be set with the merged content
-                expect(LocalForageProviderMock.localForageSet).toHaveBeenNthCalledWith(1,
+                expect(IDBKeyValProviderMock.idbKeyvalSet).toHaveBeenNthCalledWith(1,
                     '@USER_1', {
                         name: 'Tom',
                         age: 31,
@@ -91,7 +91,7 @@ describe('storage/providers/LocalForage', () => {
                         },
                     });
 
-                expect(LocalForageProviderMock.localForageSet).toHaveBeenNthCalledWith(2,
+                expect(IDBKeyValProviderMock.idbKeyvalSet).toHaveBeenNthCalledWith(2,
                     '@USER_2', {
                         name: 'Sarah',
                         age: 26,
@@ -107,31 +107,31 @@ describe('storage/providers/LocalForage', () => {
         // We're creating a Promise which we programatically control when to resolve.
         const task = createDeferredTask();
 
-        // We configure localforage.setItem to return this promise the first time it's called and to otherwise return resolved promises
-        LocalForageProviderMock.setItem = jest.fn()
+        // We configure idbKeyval.setItem to return this promise the first time it's called and to otherwise return resolved promises
+        IDBKeyValProviderMock.setItem = jest.fn()
             .mockReturnValue(Promise.resolve()) // Default behavior
             .mockReturnValueOnce(task.promise); // First call behavior
 
-        // Make 5 StorageProvider.setItem calls - this adds 5 items to the queue and starts executing the first localForage.setItem
+        // Make 5 StorageProvider.setItem calls - this adds 5 items to the queue and starts executing the first idbKeyval.setItem
         for (let i = 0; i < 5; i++) {
-            LocalForageProviderMock.setItem(`key${i}`, `value${i}`);
+            IDBKeyValProviderMock.setItem(`key${i}`, `value${i}`);
         }
 
-        // At this point,`localForage.setItem` should have been called once, but we control when it resolves, and we'll keep it unresolved.
-        // This simulates the 1st localForage.setItem taking a random time.
-        // We then call StorageProvider.clear() while the first localForage.setItem isn't completed yet.
-        LocalForageProviderMock.clear();
+        // At this point,`idbKeyval.setItem` should have been called once, but we control when it resolves, and we'll keep it unresolved.
+        // This simulates the 1st idbKeyval.setItem taking a random time.
+        // We then call StorageProvider.clear() while the first idbKeyval.setItem isn't completed yet.
+        IDBKeyValProviderMock.clear();
 
-        // Any calls that follow this would have been queued - so we don't expect more than 1 `localForage.setItem` call after the
+        // Any calls that follow this would have been queued - so we don't expect more than 1 `idbKeyval.setItem` call after the
         // first one resolves.
         task.resolve();
 
         // waitForPromisesToResolve() makes jest wait for any promises (even promises returned as the result of a promise) to resolve.
-        // If StorageProvider.clear() does not abort the queue, more localForage.setItem calls would be executed because they would
+        // If StorageProvider.clear() does not abort the queue, more idbKeyval.setItem calls would be executed because they would
         // be sitting in the setItemQueue
         return waitForPromisesToResolve().then(() => {
-            expect(LocalForageProviderMock.localForageSet).toHaveBeenCalledTimes(0);
-            expect(LocalForageProviderMock.clear).toHaveBeenCalledTimes(1);
+            expect(IDBKeyValProviderMock.idbKeyvalSet).toHaveBeenCalledTimes(0);
+            expect(IDBKeyValProviderMock.clear).toHaveBeenCalledTimes(1);
         });
     });
 });