… lib/storage/providers/LocalForage.js