<!-- If necessary, assign reviewers that know the area or changes well. Feel free to tag any additional reviewers you see fit. -->

### Details
<!-- Explanation of the change or anything fishy that is going on -->

While investigating performance issues on large accounts we found an issue where storing large amounts of data in onyx takes a very long time (E.g. merging the data from a OpenApp command).
Largely this was due to the underlying storage provider opening a new database transaction for every operation in a  multiMerge/multiSet operation.
The [idb-keyval](https://www.npmjs.com/package/idb-keyval) package optimises this, as it can handle multiple operations in one database transaction.
The good thing is, that `idb-keyval` also uses indexedDB as storage provider, so no migration is needed.

I created a performance comparison tool, and it showed how much better `idb-keyval` performs:

> Each operation has been executed with 10,000 items (depth 1), which has a size of ~2.3MB

| Operation | Onyx w/ Localforage | Onyx w/ idb-keyval | Improvement in %
|--------|--------|--------|--------|
| Single Set | 42ms | 36ms | **-14%**
| Get (collection) | 26ms | 11ms | **-58%**
| MergeCollection | 57926ms | 775ms | **-99%**
| Single Merge | 48ms | 30ms | **-38%**

<details><summary><b>How to run the performance tests yourself:<b></summary>
<p>

1. Hop on [this branch](https://github.com/margelo/react-native-onyx/tree/poc/other-databases) (its on our fork, sorry)
2. `cd example`
3. `npm i`
4. `npm start`
5. Open in your web browser `http://localhost:8080`

<img width="1624" alt="Screenshot 2023-08-03 at 16 45 41" src="https://github.com/Expensify/react-native-onyx/assets/16821682/1c5116d9-f069-48d3-ba23-f859183bac53">


</p>
</details> 

For the product the results can be seen in this video:


https://expensify.slack.com/archives/C035J5C9FAP/p1691063111819169


### Related Issues
<!-- Please replace GH_LINK with the link to the GitHub issue this Pull Request is related to -->
https://github.com/Expensify/react-native-onyx/issues/281

### Automated Tests
<!---
Most changes to Onyx should have accompanying tests. Describe the tests you added or if no tests were added an explanation about why one was not needed.
--->

### Linked PRs
<!---
Please include links to any update PRs in repos that must change their package.json version.
--->