@marcaaron 2023-08-08
>Looks great! Any idea how can we test this in product?

I would try this:

1. Run `npm build` from inside `react-native-onyx`
2. Replace the `/dist` folder in `App/node_modules/react-native-onyx`
3. Use the production API proxy in the `.env` file
4. Sign in with an account that is completely buggered and has tons of reports + policies
5. See if it improves performance