@hannojg 2023-08-10
@marcaaron You mean to delete the provider file?