@danieldoglas 2023-08-15
> maybe we can benchmark Onyx.clear as well since I know there are some race conditions around it and it's unclear how well the localforage-removeItems plugin ever actually worked

I don't think this is blocker for merging this PR since we don't know if the old one actually worked, right? 

It would be great to have this merged Today so we can see the benefits for this mostly in Guides and AM accounts