@danieldoglas 2023-08-15
Both of them return 0ms when running `Clear`

![image](https://github.com/Expensify/react-native-onyx/assets/5818999/7a785698-eaa0-4ed2-b605-38ce2fd9f068)

![image](https://github.com/Expensify/react-native-onyx/assets/5818999/8c267dea-6950-45aa-9cf8-b734f42094b3)