@hannojg 2023-08-15
Let me fix the benchmark for clearing and give numbers for that 😊  With this implementation deleting / clearing should definitely work