@hannojg 2023-08-15
Clearing seems to be a bit slower with idb-keyval, however the impact doesn't look too dramatic. I think we should be able to optimise it to be on par with local forage. However, I don't think that has too much priority at the moment / blocking this from getting merged?

<img width="1150" alt="localforage" src="https://github.com/Expensify/react-native-onyx/assets/16821682/40a74b9b-505e-470e-b947-be360a6815c5">

<img width="1150" alt="idb-keyval" src="https://github.com/Expensify/react-native-onyx/assets/16821682/df77ac94-8536-4a13-9ab8-9b3064afd04f">

(Note: for a lower item count there is almost no difference between the two implementations)