@marcaaron 2023-09-01
>While investigating performance issues on large accounts we found an issue where storing large amounts of data in onyx takes a very long time (E.g. merging the data from a OpenApp command).
>Largely this was due to the underlying storage provider opening a new database transaction for every operation in a multiMerge/multiSet operation.

Are we sure this is faster than the previous solution of setting one value at a time synchronously? Did we verify that things improved on a large account? Which account did we test with?

>The [idb-keyval](https://www.npmjs.com/package/idb-keyval) package optimises this, as it can handle multiple operations in one database transaction.

Is is possible we can ask it to do too many at once? I am seeing various weird logs like:


```
Error: TimeoutError: Transaction timed out due to inactivity.. onyxMethod: Ae ~~ parameters: '' userAgent: 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/116.0.0.0 Safari/537.36 Edg/116.0.1938.54'
```