@Szymon20000 2023-09-04
I will try but my focus this week is on improving chat switching and batching updates.